tests/wpt/web-platform-tests/css/css-contain/content-visibility/content-visibility-with-top-layer-006.html

<!doctype html>
<meta charset="utf8">
<title>CSS Content Visibility: offscreen c-v auto content is relevant when in top layer</title>
<link rel="author" title="Martin Robinson" href="mailto:mrobinson@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert" content="offscreen c-v auto content is relevant when in top layer">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
#inner {
    content-visibility: auto;
    contain-intrinsic-size: 100px 100px;
}
</style>

<dialog id="dialog">
    <div id="spacer" style="height: 100000px;"></div>
    <div id="inner">
        <div style="height: 200px; width: 200px;">content</div>
    </div>
</dialog>

<script>
test(() => {
    dialog.show();

    assert_equals(inner.getBoundingClientRect().height, 100);

    dialog.close();
    dialog.showModal();

    // Even though the element with `content-visibility: auto` is offscreen
    // it should be considered relevant content because it is now in the top
    // layer.
    assert_equals(inner.getBoundingClientRect().height, 200);
});
</script>
